requirements.txt
requests~=2.26.0
beautifulsoup4~=4.9.3
reportlab~=3.6.1
numpy~=1.21.2
Pillow~=8.3.1